Add files via upload
Group project to create a library application that allowed for loans of books and DVDs while also reading and recording data from .csv files. The application uses inheritance, overloading, and overriding as well as an object-oriented approach. My contribution was the majority of the methods added as well as the implementation of inheritance and creation of the parent abstract class and its children. I was also heavily involved with testing and data validation. We achieved a grade of 94% for this project.